Rework tools page.
1-column layout. Ordering of the tools roughly is: showcase / highlight,
in pipeline, the oldies. Allow referencing associated publication in
full instead of just saying "Publication". Order per tool data as: info,
publication (if any), link to code and then the status. Improved info of
some tools.

We need better screenshots for most of the tools and should probably
adjust the whitespace a bit.

close #15

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>